Release 0.0.2

+++ src/main/java/lt/karijotas/microblogging/api/PostController.java
@@ -1,16 +1,29 @@
 package lt.karijotas.microblogging.api;
 
+import lt.karijotas.microblogging.dao.CommentRepository;
 import lt.karijotas.microblogging.exception.BlogValidationExeption;
+import lt.karijotas.microblogging.model.Comment;
 import lt.karijotas.microblogging.model.Post;
 import lt.karijotas.microblogging.model.dto.PostDto;
 import lt.karijotas.microblogging.model.dto.PostEntityDto;
+import lt.karijotas.microblogging.service.BloggerService;
+import lt.karijotas.microblogging.service.CommentService;
 import lt.karijotas.microblogging.service.PostService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
 
 import javax.validation.Valid;
 import java.util.List;
@@ -24,50 +37,81 @@ import static org.springframework.http.ResponseEntity.ok;
 public class PostController {
     private final Logger logger = LoggerFactory.getLogger(PostController.class);
     private final PostService postService;
+    private final BloggerService BloggerService;
+    private final CommentService commentService;
+    private final CommentRepository commentRepository;
 
-    public PostController(PostService postService) {
+    public PostController(PostService postService, BloggerService bloggerService, CommentService commentService,
+                          CommentRepository commentRepository) {
         this.postService = postService;
+        this.BloggerService = bloggerService;
+        this.commentService = commentService;
+        this.commentRepository = commentRepository;
     }
 
-    @GetMapping(produces = {MediaType.APPLICATION_JSON_VALUE,})
+    @GetMapping(produces = {MediaType.APPLICATION_JSON_VALUE})
     @ResponseBody
     public List<Post> getAll() {
         return postService.getAll();
     }
 
-    @GetMapping(value = "/{postId}", produces = {MediaType.APPLICATION_JSON_VALUE,})
-    public ResponseEntity<PostEntityDto> getPost(@PathVariable Long postId) {
-        var postOptional = postService.getById(postId);
+    @GetMapping(value = "/user", produces = {MediaType.APPLICATION_JSON_VALUE})
+    public ResponseEntity<List<Post>> getAllByCurrentUser(@AuthenticationPrincipal UserDetails userDetails) {
+        if (userDetails != null) {
+            String username = userDetails.getUsername();
+            Long id = BloggerService.findByUserName(username).getId();
+            return ResponseEntity.ok(postService.getAllByCurrentAuthor(id));
+        }
+        return null;
+    }
 
-        var responseEntity = postOptional
-                .map(post -> ok(toPostEntityDto(post)))
-                .orElseGet(() -> ResponseEntity.notFound().build());
+    @GetMapping(value = "/user/{userId}", produces = {MediaType.APPLICATION_JSON_VALUE})
+    @ResponseBody
+    public List<Post> getAllByUserId(@PathVariable Long userId) {
+        return postService.getAllByAuthor(userId);
+    }
 
-        return responseEntity;
+    @GetMapping(value = "/{postId}", produces = {MediaType.APPLICATION_JSON_VALUE})
+    @ResponseBody
+    public Post getPost(@PathVariable Long postId) {
+        var postOptional = postService.getById(postId).orElseThrow(
+                () -> new BlogValidationExeption("Post doesn't exist", "id", "Post doesn't exist", postId.toString()));
+        return postOptional;
     }
 
     @PostMapping(produces = {MediaType.APPLICATION_JSON_VALUE})
     public ResponseEntity<PostEntityDto> create(@Valid @RequestBody PostEntityDto postEntityDto) {
-
         var createdPost = postService.create(postEntityDto);
         return ok(toPostEntityDto(createdPost));
     }
 
     @DeleteMapping("/{postId}")
-    public ResponseEntity<Void> deletePost(@PathVariable Long postId) {
-        logger.info("Attempt to delete Post by id: {}", postId);
-        boolean deleted = postService.deleteById(postId);
-        if (deleted) {
-            return ResponseEntity.noContent().build();
-        } else {
-            return ResponseEntity.notFound().build();
+    public ResponseEntity<Void> deletePost(@PathVariable Long postId, @AuthenticationPrincipal UserDetails userDetails) {
+        if (userDetails != null) {
+            Long id = BloggerService.findByUserName(userDetails.getUsername()).getId();
+            if (postService.validateOwnership(id, postId)) {
+                logger.info("Attempt to delete Post by id: {}", postId);
+                List<Comment> comments = commentService.getAllByPostId(postId);
+                logger.info("Deleting {} comments from the post", comments.size());
+                commentRepository.deleteAll(comments);
+                comments.clear();
+                boolean deleted = postService.deleteById(postId);
+                return (deleted ? ResponseEntity.noContent() : ResponseEntity.notFound()).build();
+            }
         }
+        return ResponseEntity.notFound().build();
     }
 
     @PatchMapping("/{postId}")
-    public ResponseEntity<PostDto> updatePost(@PathVariable Long postId, @Valid @RequestBody PostDto postDto) {
-        var updated = postService.update(toPost(postDto), postId);
-        return ok(toPostDto(updated));
+    public ResponseEntity<PostDto> updatePost(@PathVariable Long postId, @Valid @RequestBody PostDto postDto, @AuthenticationPrincipal UserDetails userDetails) {
+        if (userDetails != null) {
+            Long id = BloggerService.findByUserName(userDetails.getUsername()).getId();
+            if (postService.validateOwnership(id, postId)) {
+                var updated = postService.update(toPost(postDto), postId);
+                return ok(toPostDto(updated));
+            }
+        }
+        return ResponseEntity.notFound().build();
     }
 
     @GetMapping(value = "/{postId}/wordcount", produces = MediaType.APPLICATION_JSON_VALUE)
@@ -79,12 +123,11 @@ public class PostController {
     }
 
 
-    @GetMapping(value = "/{postId}/wordcount/{limit}", produces = {MediaType.APPLICATION_JSON_VALUE,})
+    @GetMapping(value = "/{postId}/wordcount/{limit}", produces = {MediaType.APPLICATION_JSON_VALUE})
     public ResponseEntity<Map<String, Long>> mostUsedWords(@PathVariable Long postId, @PathVariable Long limit) {
         Post post = postService.getById(postId)
                 .orElseThrow(() -> new BlogValidationExeption("Post doesn't exist", "id", "Post doesn't exist", postId.toString()));
         Map<String, Long> mostUsedWords = postService.mostUsedWords(post, limit);
         return ResponseEntity.ok().body(mostUsedWords);
     }
-
 }
